refactor(mobile): 역 시설 카드 컴포넌트 분리
