+++ apps/mobile/lib/station_search.dart
@@ -19,8 +19,9 @@ import 'features/stations/application/station_search_controller.dart';
 import 'features/stations/domain/station_line.dart';
 import 'features/stations/domain/station_models.dart';
 import 'features/stations/domain/station_repositories.dart';
+import 'features/stations/presentation/station_detail_info_row.dart';
 import 'features/stations/presentation/station_exit_card.dart';
-import 'features/stations/presentation/station_facility_detail_screen.dart';
+import 'features/stations/presentation/station_facility_card.dart';
 import 'features/stations/presentation/station_info_basis_disclosure.dart';
 import 'features/stations/presentation/station_line_badges.dart';
 import 'features/stations/presentation/station_recent_search_section.dart';
@@ -995,7 +996,7 @@ class _StationDetailContent extends StatelessWidget {
         const _StationDetailEmptyMessage(message: '시설 안내를 준비 중이에요.')
       else
         for (final facility in facilities)
-          _StationFacilityCard(
+          StationFacilityCard(
             facility: facility,
             station: detail,
             onReportTap: () => _openFacilityReport(context, facility),
@@ -1734,7 +1735,7 @@ class _StationInternalRouteGuidance extends StatelessWidget {
       InternalRouteViewStatus.loading => Semantics(
         label: '역 안 이동 순서 불러오는 중',
         liveRegion: true,
-        child: const _StationDetailInfoRow(
+        child: const StationDetailInfoRow(
           icon: Icons.sync,
           text: '역 안 이동 순서를 불러오는 중입니다.',
         ),
@@ -1742,7 +1743,7 @@ class _StationInternalRouteGuidance extends StatelessWidget {
       InternalRouteViewStatus.failure => Semantics(
         label: state.message,
         liveRegion: true,
-        child: _StationDetailInfoRow(
+        child: StationDetailInfoRow(
           icon: Icons.error_outline,
           text: state.message,
         ),
@@ -1777,7 +1778,7 @@ class _StationInternalRouteResultCard extends StatelessWidget {
           child: Column(
             crossAxisAlignment: CrossAxisAlignment.start,
             children: [
-              _StationDetailInfoRow(
+              StationDetailInfoRow(
                 icon: result.statusIcon,
                 text: result.statusLabel,
               ),
@@ -1802,7 +1803,7 @@ class _StationInternalRouteResultCard extends StatelessWidget {
               if (result.warnings.isNotEmpty) ...[
                 const SizedBox(height: 10),
                 for (final warning in result.warnings)
-                  _StationDetailInfoRow(
+                  StationDetailInfoRow(
                     icon: Icons.warning_amber,
                     text: warning.message,
                   ),
@@ -1869,33 +1870,6 @@ class _StationInternalRouteStepTile extends StatelessWidget {
   }
 }
 
-class _StationDetailInfoRow extends StatelessWidget {
-  const _StationDetailInfoRow({required this.icon, required this.text});
-
-  final IconData icon;
-  final String text;
-
-  @override
-  Widget build(BuildContext context) {
-    return Row(
-      children: [
-        Icon(icon, size: 22, color: EasySubwayAccessibleColors.primary),
-        const SizedBox(width: 8),
-        Expanded(
-          child: Text(
-            text,
-            style: Theme.of(context).textTheme.bodyLarge?.copyWith(
-              color: EasySubwayAccessibleColors.secondaryText,
-              fontWeight: FontWeight.w700,
-              height: 1.3,
-            ),
-          ),
-        ),
-      ],
-    );
-  }
-}
-
 class _StationDetailSectionTitle extends StatelessWidget {
   const _StationDetailSectionTitle({required this.title});
 
@@ -1930,152 +1904,6 @@ class _StationDetailEmptyMessage extends StatelessWidget {
         color: EasySubwayAccessibleColors.secondaryText,
         fontWeight: FontWeight.w700,
         height: 1.35,
-      ),
-    );
-  }
-}
-
-class _StationFacilityCard extends StatelessWidget {
-  const _StationFacilityCard({
-    required this.facility,
-    required this.station,
-    required this.onReportTap,
-  });
-
-  final StationFacilityInfo facility;
-  final StationDetail station;
-  final VoidCallback onReportTap;
-
-  @override
-  Widget build(BuildContext context) {
-    final textTheme = Theme.of(context).textTheme;
-
-    return Semantics(
-      container: true,
-      explicitChildNodes: true,
-      label: facility.semanticLabel,
-      button: true,
-      onTap: () => _openFacilityDetail(context),
-      child: Card(
-        margin: const EdgeInsets.only(bottom: 12),
-        color: Colors.white,
-        elevation: 0,
-        shape: const RoundedRectangleBorder(
-          borderRadius: _stationDetailFacilityCardRadius,
-          side: BorderSide(color: EasySubwayAccessibleColors.line),
-        ),
-        child: InkWell(
-          key: Key('stationFacilityCard-${facility.id}'),
-          borderRadius: _stationDetailFacilityCardRadius,
-          onTap: () => _openFacilityDetail(context),
-          child: Padding(
-            padding: const EdgeInsets.all(16),
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                Text(
-                  facility.name,
-                  style: textTheme.titleMedium?.copyWith(
-                    color: EasySubwayAccessibleColors.text,
-                    fontWeight: FontWeight.w700,
-                    height: 1.25,
-                  ),
-                ),
-                // 정상 시설은 필 없이 조용히 표시하고, 문제(고장·공사)일 때만
-                // 상태 필 하나만 노출한다. 시설 종류는 이름에 이미 포함되고,
-                // '이용 가능'='정상' 중복과 종류 필은 제거한다.
-                if (facility.needsAttention) ...[
-                  const SizedBox(height: 10),
-                  _StationDetailTextPill(text: facility.statusTitle),
-                ],
-                const SizedBox(height: 12),
-                _StationDetailInfoRow(
-                  icon: Icons.place_outlined,
-                  text: facility.locationLabel,
-                ),
-                const SizedBox(height: 6),
-                _StationDetailInfoRow(
-                  icon: Icons.event_available,
-                  text: facility.updatedLabel,
-                ),
-                const SizedBox(height: 8),
-                // 상용 리스트 항목처럼 카드 전체 탭으로 상세에 들어가고(우측 ›로
-                // 암시), 보조 액션 '시설 알려주기'는 텍스트 버튼 수준으로 낮춘다.
-                // 카드 탭과 중복되던 '상세 보기' 텍스트는 제거한다.
-                Row(
-                  children: [
-                    Semantics(
-                      container: true,
-                      label: '${facility.name} 시설 알려주기',
-                      button: true,
-                      onTap: onReportTap,
-                      child: ExcludeSemantics(
-                        child: TextButton.icon(
-                          key: Key('facilityReportButton-${facility.id}'),
-                          onPressed: onReportTap,
-                          style: TextButton.styleFrom(
-                            padding: const EdgeInsets.symmetric(
-                              horizontal: 8,
-                              vertical: 8,
-                            ),
-                          ),
-                          icon: const Icon(Icons.report_outlined, size: 20),
-                          label: const Text('시설 알려주기'),
-                        ),
-                      ),
-                    ),
-                    const Spacer(),
-                    const Icon(
-                      Icons.chevron_right,
-                      color: EasySubwayAccessibleColors.mutedText,
-                    ),
-                  ],
-                ),
-              ],
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-
-  void _openFacilityDetail(BuildContext context) {
-    Navigator.of(context).push(
-      MaterialPageRoute<void>(
-        builder: (_) => FacilityDetailScreen(
-          station: station,
-          facility: facility,
-          onReportTap: onReportTap,
-        ),
-      ),
-    );
-  }
-}
-
-class _StationDetailTextPill extends StatelessWidget {
-  const _StationDetailTextPill({required this.text});
-
-  final String text;
-
-  @override
-  Widget build(BuildContext context) {
-    // 유형·품질 라벨은 상태 의미가 없으므로 틴트 필 대신 중립 아웃라인으로.
-    return Container(
-      decoration: BoxDecoration(
-        color: EasySubwayAccessibleColors.surface,
-        borderRadius: _stationDetailInfoCardRadius,
-        border: Border.all(color: EasySubwayAccessibleColors.line),
-      ),
-      child: Padding(
-        padding: const EdgeInsets.symmetric(horizontal: 12, vertical: 8),
-        child: Text(
-          text,
-          style: Theme.of(context).textTheme.bodyLarge?.copyWith(
-            color: EasySubwayAccessibleColors.secondaryText,
-            fontWeight: FontWeight.w700,
-            height: 1.2,
-          ),
-        ),
       ),
     );
   }
